Particle re adjustment

+++ src/interfaces/Drawable.java
@@ -4,5 +4,5 @@ import java.awt.Graphics;
 
 public interface Drawable {
 	public void draw(Graphics g);
-	public void update(long dt);
+	public void update(double dt);
 }
